backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

google-cloud-storage==2.10.0
google-cloud-documentai==2.20.1
google-cloud-vision==3.4.4

langchain==0.0.350
langchain-anthropic==0.0.2
langchain-pinecone==0.0.1
pinecone-client==2.2.4
langgraph==0.0.20
langsmith==0.0.69

PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
openpyxl==3.1.2

pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2

pytest==7.4.3
pytest-asyncio==0.21.1

# fastapi
# uvicorn[standard]
# google-cloud-storage
# google-cloud-firestore
# google-cloud-vision
# google-cloud-documentai
# pinecone
# langchain
# langchain-community
# langchain-core
# langchain-anthropic
# langchain-google-vertexai
# langchain-tavily
# langchain-pinecone
# langgraph
# langsmith
# anthropic
# python-dotenv
# tiktoken
# unstructured[all] 
# pydantic
# markitdown
# pandas
# markdown
# PyPDF2
# python-docx
# pytesseract
# huggingface-hub 
# python-multipart 
# typing-extensions
